+++ src/modules/scoreClass.js
